28a_tkinter/38_scaling_image/requirements.txt
pillow==12.0.0
customtkinter==5.2.2
